@@ -1,6 +1,8 @@
 import { Component, inject } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { Router } from '@angular/router';
 import { AuthService } from '@modules/auth/services/auth.service';
+import { CookieService } from 'ngx-cookie-service';
 
 @Component({
   selector: 'app-auth-page',
@@ -8,6 +10,7 @@ import { AuthService } from '@modules/auth/services/auth.service';
   styleUrls: ['./auth-page.component.css']
 })
 export class AuthPageComponent {
+  errorSession:boolean = false;
 
   loginForm = new FormGroup({
     email: new FormControl('', [
@@ -21,9 +24,18 @@ export class AuthPageComponent {
     ])
   });
   private _authService = inject(AuthService);
+  private router = inject(Router);
 
   sendCredential(): void {
     const { email, password } = this.loginForm.value as {email:string,password:string};
-    this._authService.sendCredentials(email, password);
+    this._authService.sendCredentials(email, password).subscribe({
+      next: logInfo =>{
+        this.router.navigate(['/','tracks']);
+        this.errorSession = false;
+      },
+      error: err=>{
+        this.errorSession = true;
+      }
+    });
   }
 }
